Se creo la estructura de datos ListaCola y se agrego un paquete llamado res.interfazDB donde se crearan las clases para manejar los datos de las interfaces. Se reajustaron valores en las conexiones a sql.

+++ src/res/Producto.java
@@ -17,7 +17,7 @@ public class Producto extends AbarrotesElemento{
     private String um;
     
     public Producto(int id, String nombre, int existencia, int stock,
-            int precio, String um){
+            float precio, String um){
         this.id = id;
         this.nombre = nombre;
         this.existencia = existencia;
